backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
